add grid2 type and replace use of index.h in eik with it
The idea here is that whether the grid is row or column major will
become a run-time parameter. The next thing we want to do is fix all
the different indexing and offsetting stuff in eik to match the order
of the grid, and to replace the arrays that are computed statically
with ones that are computed at runtime.